Refactor switchting between enter and exit indoor map
Previously, an extra reference to the empty IndoorView GameObject was
needed. This is not the case anymore. When entering the indoor map, all
floors except the current one will be hidden. When exiting the indoor
map, all floors will be hidded and the outside model will be shown. The
map is not entered by default (i.e., the outside model is shown by
default).