Improve LocationPin for VR a LOT
LocationPins now look much better and are properly scaled.
'Zurück' and 'OK' buttons work both in VR and Fallback by pulling the
Vive trigger.
The LocationPin info can be shown by grabbing one of the pins with the
Vive controller trigger (doesn't work in fallback because of SteamVR bug
with the simulated hand).

Styling of the pin itself is still ToDo.