Create LaserPointer for activating LocationPins
A ray is emitted from the preffered hand that will have its endpoint
when it is hitting any colliders. In that case, it moves the hover
Transform of the SteamVR hand to the end of the ray, which *should* let
the hand interact with any remote SteamVR Interactables that are placed
at the end of the ray.

Closes #35.