Fix laserpointer being shown at start
The laser pointer is not hidden at startup because it is not a child of
any [SUBTASK] GO, meaning its OnEnable are called directly with the
first subtask (user setup).

This commit fixes the issue by providing a new method Show() and
calling it with false in the laser pointer's own Start() method.