Fix #43: Trim last empty line before parsing sequences
This is what commit 632c187b3a4fcd07b0dc9ccf9ef98962c519ddf1 should have
fixed already. However, the last empty line was only skipped for
parsing, the size of the sequence list stayed the same (meaning it was 1
element too big, leading to a NullReference when trying to select that
line as active sequence).

This commit fixes #43 by first trimming trailing whitespace and THEN
parsing. Note that empty lines in the middle of the file are still NOT
supported (and why should it...).