Fix Pointer activation by using LaserPointer Events
When changing hit colliders, the laser emits events now (passing both
source and destination colliders). These are used in the case handling
of LocationPins to determine in which cases to show the info or not.

The reason this is done in such a complex way is because before the
Hover events of the SteamVR hands were used. The laser moved the Hover
transform towards the hit point on the collider. Unfortunately, the
buttons colliders were not connected to the Pin's colliders, resulting
in a premature Hide(). Also, adding a thin box collider for the
background alone did not help, because the button's box colliders would
go beyond the background collider, leading to another Hide() event.

This commit ACTUALLY fixes #41.