Sort user folders by ordinal value, not alphabetic
A third-party string comparer is used to sort the already existing user
folder names when creating a new one. Before doing this, we would create
multiple 'user_10_...' dirs because alphabetically 10 comes after 1 and
before 2, so it was sorted early in the list and the last element was
always 'user_9_...'. With the new comparer, 'user_10_...' is now
properly put at the end of the list and afterwards correctly incremented
to 11.

Fixes https://gitlab.com/mowalle/megamap/issues/3.