Fix balls floating in air
We used the balls' sphere collider radius for positioning inside the
room on the y axis. However, the collider is not scaled by the balls'
transform scale, leading to a bigger y offset. This commit makes it so
we scale the radius before we add it to the ball position.