Extend task handling for tutorial levels
The task switcher can have multiple tutorial levels assigned that use a
tutorial condition previously set up in the condition switcher. After
all tutorials are completed, the experimenter can start the acutal
experiment or repeat the tutorials if the user so desires.

No experiment data is recorded during the tutorials. However, HMD and
hand positions as well as the log are still recorded for completeness.

Note that separate tutorial level layouts for indoor maps will be added
in future commits.

Closes https://github.com/Mowalle/megamap/issues/7.