Fix task still being repeatable when switching conditions
Before this commit, the number of finished tasks was reset BEFORE the
next sequence was generated. However, since for remembering the last
active task within NextSequence() the CurrenTaskIndex was used, this let
to the wrong task being remembered as the last task (i.e., always the
task at startOffset was remembered). This had the effect that the start
of the previous sequence was not repeated rather than the end of the
previous sequence.

The fix is to reset the number of finished tasks AFTER the next sequence
is generated. This commit fixes https://github.com/Mowalle/megamap/issues/9.