Use adaptive height instead of fixed for 'High' condition
If the Megamap's HeightMode is set to 'fixed',  the heightOffset
determines the maps height from the floor. When it's set to 'adaptive',
the heightOffset determines the height below the user's head (the
tracked HMD position). The height offset will be applied to the Megamap
when calling the Show() method (thus, changing these values while the
map is shown has no effect yet).

In 2D-mode, these values are ignored (the map placement is determined
entirely by the referenced 2D transform).

Closes #22.