Allow task to be marked as skipped by instructor
When experiment tasks are running, the instructor can press the 'X' key
to toggle the skip state of the current task. By doing this, a message
will be shown in the instructor UI that will show the skip state of the
current task, as well as the time until the task was marked as skipped.

This can be used if a user gets stuck on a task (for example, if he
doesn't find the correct number of balls). Note that the task itself
will not actually be skipped; the instructor has to go into 2D fallback
mode and finish the task manually. Afterwards, (s)he has to enable VR
mode again.

See #1.